ReferalApp/requirements.txt
Django==3.2.0
psycopg2-binary==2.9.1
django-cors-headers==3.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0